requirements.txt
python-openstackclient
diagrams